Реализовал по мотивам вебинара контейнер перед компонентом ChatForm.
создал новую функцию onSendMessage из addMessage с назначенным параметром id
